% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_JTRAP_Marsh.R
\docType{data}
\name{Data_JTRAP_Marsh}
\alias{Data_JTRAP_Marsh}
\title{Example rotary screw trap data, Marsh Creek, ID, 2005-2015}
\format{Data is formatted and ready for \code{MYTSBE()}. Originally a .CSV file}
\usage{
Data_JTRAP_Marsh
}
\description{
This is an example data of juvenile salmonids at a rotary screw trap on Marsh Creek, ID, 2005-2015.
}
\keyword{Chinook}
\keyword{and}
\keyword{data}
\keyword{juvenile}
\keyword{steelhead}
\keyword{trap}

