% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_PTAGIS_Marsh_Chin.R
\docType{data}
\name{Data_PTAGIS_Marsh_Chin}
\alias{Data_PTAGIS_Marsh_Chin}
\title{Example juvenile Chinook PTAGIS data, Marsh Creek, ID, 2005-2015}
\format{Data is formatted and ready for \code{MYTSBE()}.}
\usage{
Data_PTAGIS_Marsh_Chin
}
\description{
This is an example PTAGIS data for Chinook at Marsh Creek,ID, 2005-2015.
}
\keyword{Chinook}
\keyword{Creek}
\keyword{Marsh}
\keyword{PTAGIS}
\keyword{data}
\keyword{example}
\keyword{for}

