% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_PTAGIS_Marsh_Steel.R
\docType{data}
\name{Data_PTAGIS_Marsh_Steel}
\alias{Data_PTAGIS_Marsh_Steel}
\title{Example juvenile Steelhead PTAGIS data, Marsh Creek, ID, 2005-2015}
\format{Data is formatted and ready for \code{MYTSBE()}.}
\usage{
Data_PTAGIS_Marsh_Steel
}
\description{
This is an example PTAGIS data  for Steelhead at Marsh Creek,ID, 2005-2015.
}
\keyword{Creek}
\keyword{Marsh}
\keyword{PTAGIS}
\keyword{Steelhead}
\keyword{data}
\keyword{example}
\keyword{for}

