% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYTSBE_Cohort_Fry.R
\name{MYTSBE_Cohort_Fry}
\alias{MYTSBE_Cohort_Fry}
\title{Multi-year time stratified Bayesian estimator (cohort summary)}
\usage{
MYTSBE_Cohort_Fry(data, effort.cor = FALSE, sel.years = currentyear,
  strata.op.min = 1, smolt.parr.date = "07-01",
  parr.presmolt.date = "09-01", species = "", trap.name = "",
  den.plot = TRUE, trace.plot = TRUE, strata.length = s.length,
  burnin = 1e+05, chains = 3, iterations = 4e+05, thin = 100,
  boot = 5000, rm.bad.GR = FALSE, model.params = c("p", "U", "etaP1",
  "etaU1", "sigmaU", "sigmaP"))
}
\arguments{
\item{data}{capture-mark-recapture data frame}

\item{effort.cor}{expands the number of unmarked fish captured by a sampling effort}

\item{sel.years}{selected year(s) to compute abundance estimates}

\item{strata.op.min}{minimum number of years data need to have been collected in a stratum to be included in summary}

\item{smolt.parr.date}{"MM-DD" date to partition smolt life stage}

\item{parr.presmolt.date}{"MM-DD" date to partition parr life stage}

\item{species}{character string used for titles and descriptions of reports}

\item{trap.name}{character string used for  titles and descriptions of reports}

\item{den.plot}{return density plots of MCMC chains}

\item{trace.plot}{return trace plots of MCMC chains}

\item{strata.length}{number of days in strata}

\item{burnin}{number of initial MCMC chain iterations to be discarded}

\item{chains}{number of MCMC chains (>1)}

\item{iterations}{number of MCMC iterations per chain}

\item{thin}{thin rate}

\item{boot}{number of boot strap iterations to calculate yearly and life stage abundance estimates}

\item{rm.bad.GR}{remove U posterior distributions for strata with Gelman-Rubins test statistic >1.1 from yearly summary}

\item{model.params}{parameters to be returned from MCMC simulation}
}
\description{
MYTSBE utilizes hierarchical Bayesian methods to estimate abundances from capture-mark-recapture (CMR) data using a temporally-stratified Lincoln-Petersen estimator. The between-year hierarchical
structures allows for annually recurring species characteristics to be incorporated into capture probabilities and abundance estimates.

Summary output is by cohort year.

Peer reviewed publication can be found at ...
}

